项目1-英文翻译/sound_tech_AI_translate/requirements.txt
deepseek
fastapi
uvicorn
python-docx